requirements.txt
autocommand==2.2.2
backports.tarfile==1.2.0
charset-normalizer==3.4.4
cheroot==11.0.0
CherryPy==18.10.0
et_xmlfile==2.0.0
jaraco.collections==5.2.1
jaraco.context==6.0.1
jaraco.functools==4.3.0
jaraco.text==4.0.0
more-itertools==10.8.0
numpy==2.3.4
openpyxl==3.1.5
pandas==2.3.3
passlib==1.7.4
pillow==12.0.0
portend==3.2.1
psycopg2==2.9.11
PyJWT==2.10.1
python-dateutil==2.9.0.post0
pytz==2025.2
reportlab==4.4.4
six==1.17.0
tempora==5.8.1
tzdata==2025.2
zc.lockfile==4.0
